// client/src/app/core/models/card-type.ts
export enum CardType {
  UsedCard = 'UsedCard',
  FromCommonDeck = 'FromCommonDeck',
  Plus = 'Plus',
  Minus = 'Minus',
  PlusMinus = 'PlusMinus',
  PlusMinus1 = 'PlusMinus1',
  PlusMinus2 = 'PlusMinus2',
  PlusMinus3 = 'PlusMinus3',
  PlusMinus4 = 'PlusMinus4',
  PlusMinus5 = 'PlusMinus5',
  PlusMinus6 = 'PlusMinus6',
  OneOrTwoPlusMinus = 'OneOrTwoPlusMinus',
  ThreeOrFourPlusMinus = 'ThreeOrFourPlusMinus',
  FiveOrSixPlusMinus = 'FiveOrSixPlusMinus',
  AnyValue = 'AnyValue',
}